:bug: Only a single pathway to connect a PerseverantConsumer.
Connecting in the NewPerseverantConsumer method and also connecting if necessary in StartListening could lead to a failure to receive any messages, with the message handler function assigned to a dead Kafka connection.

Now, PerseverantConsumers are connected only on demand. This introduces the possibility that a consumer healthcheck may fail momentarily on startup of an essentially healthy service, but this is a transient event that strictly accurately represents the state of the connection.